Rename the examples module to generation
