--- docs/api/pydantic_evals/examples.md
@@ -1,3 +1,0 @@
-# `pydantic_evals.examples`
-
-::: pydantic_evals.examples
