+++ docs/api/pydantic_evals/generation.md
@@ -1,0 +1,3 @@
+# `pydantic_evals.generation`
+
+::: pydantic_evals.generation
